ainTempCaminoValue as long instead of int

+++ arduino/riscaldamento/progetto/sketches/_20170117_Override_Termostato_Ambiente/State.h
@@ -58,7 +58,7 @@ class State {
   unsigned long outOverrideTermoAmbienteAccu_On;
   unsigned long outOverrideTermoAmbienteAccu_Off;
   
-  int   ainTempCaminoValue;
+  long  ainTempCaminoValue;
   float ainTempCaminoValueCentigradi;
 };
